+++ crates/assistant2/Cargo.toml
@@ -46,7 +46,6 @@ itertools.workspace = true
 language.workspace = true
 language_model.workspace = true
 language_model_selector.workspace = true
-language_models.workspace = true
 log.workspace = true
 lsp.workspace = true
 markdown.workspace = true
