+++ crates/assistant_context_editor/Cargo.toml
@@ -30,7 +30,6 @@ indexed_docs.workspace = true
 language.workspace = true
 language_model.workspace = true
 language_model_selector.workspace = true
-language_models.workspace = true
 log.workspace = true
 multi_buffer.workspace = true
 open_ai.workspace = true
