+++ crates/language_models/Cargo.toml
@@ -41,7 +41,6 @@ settings.workspace = true
 smol.workspace = true
 strum.workspace = true
 theme.workspace = true
-thiserror.workspace = true
 tiktoken-rs.workspace = true
 ui.workspace = true
 util.workspace = true
