+++ crates/zed/src/main.rs
@@ -436,7 +436,7 @@ fn main() {
             cx,
         );
         supermaven::init(app_state.client.clone(), cx);
-        language_model::init(cx);
+        language_model::init(app_state.client.clone(), cx);
         language_models::init(
             app_state.user_store.clone(),
             app_state.client.clone(),
