+++ crates/zed/src/zed.rs
@@ -4237,7 +4237,7 @@ mod tests {
                 cx,
             );
             image_viewer::init(cx);
-            language_model::init(cx);
+            language_model::init(app_state.client.clone(), cx);
             language_models::init(
                 app_state.user_store.clone(),
                 app_state.client.clone(),
