+++ crates/zeta/Cargo.toml
@@ -33,7 +33,7 @@ http_client.workspace = true
 indoc.workspace = true
 inline_completion.workspace = true
 language.workspace = true
-language_models.workspace = true
+language_model.workspace = true
 log.workspace = true
 menu.workspace = true
 migrator.workspace = true
